Add files via upload
The article "Hunt and Undernourishment," which was posted on "Our World in Data,"  website served as the foundation for the database that was developed and the database that was analyzed. The article was written by Max Roser and Hannah Ritchie. The database is primarily based on the second article's second section's final subcategory- "undernourishment." I focused on the data from the graphs in the "Depth of food deficit" category.